requirements.txt
langchain
langchain-openai
langchain-core
openai
python-dotenv
